Conflicts:
	WebApp/src/main/java/eu/komanda30/kupra/controllers/menu/MenuController.java
	WebApp/src/main/java/eu/komanda30/kupra/controllers/menu/MenuItemForm.java
	WebApp/src/main/java/eu/komanda30/kupra/entity/Menu.java
	WebApp/src/main/resources/messages/messages_en.properties
	WebApp/src/main/resources/messages/messages_lt.properties
	WebApp/src/main/resources/messages/validation_en.properties
	WebApp/src/main/resources/messages/validation_lt.properties
	WebApp/src/main/webapp/WEB-INF/templates/accessDeniedError.html
	WebApp/src/main/webapp/WEB-INF/templates/authenticationError.html
	WebApp/src/main/webapp/WEB-INF/templates/genericError.html
	WebApp/src/main/webapp/WEB-INF/templates/layout.html
	WebApp/src/main/webapp/WEB-INF/templates/login.html
	WebApp/src/main/webapp/WEB-INF/templates/pageNotFoundError.html
	WebApp/src/main/webapp/WEB-INF/templates/popups/newProduct.html
	WebApp/src/main/webapp/WEB-INF/templates/systemError.html
	WebApp/src/main/webapp/WEB-INF/templates/unit-list.html
	WebApp/src/main/webapp/WEB-INF/templates/unsupportedError.html

+++ WebApp/src/main/java/eu/komanda30/kupra/controllers/reciperead/RecipeProductUnit.java
@@ -1,12 +1,11 @@
 package eu.komanda30.kupra.controllers.reciperead;
 
-/**
- * Created by Ignas on 12/10/2014.
- */
+import java.math.BigDecimal;
+
 public class RecipeProductUnit {
     private String name;
     private String unit;
-    private Double quantity;
+    private BigDecimal quantity;
 
     public String getName() {
         return name;
@@ -24,11 +23,11 @@ public class RecipeProductUnit {
         this.unit = unit;
     }
 
-    public Double getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
 }
