+++ WebApp/src/main/java/eu/komanda30/kupra/entity/RecipeProduct.java
@@ -1,5 +1,7 @@
 package eu.komanda30.kupra.entity;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -9,9 +11,6 @@ import javax.persistence.ManyToOne;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 
-/**
- * Created by Ignas on 12/3/2014.
- */
 @Entity
 @Table(name="recipe_product")
 @SequenceGenerator(
@@ -32,13 +31,13 @@ public class RecipeProduct {
     @JoinColumn(name="product_id")
     private Product product;
 
-    private Double quantity;
+    private BigDecimal quantity;
 
-    public Double getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
